Cuboid -> Box, re-add cube, move shapes to their own file
