+++ crates/bevy_render/src/mesh/mod.rs
@@ -1,4 +1,6 @@
 #[allow(clippy::module_inception)]
 mod mesh;
+/// Generation for some primitive shape meshes.
+pub mod shape;
 
 pub use mesh::*;
