+++ examples/ios/src/lib.rs
@@ -25,15 +25,15 @@ fn setup(
     commands
         // plane
         .spawn(PbrBundle {
-            mesh: meshes.add(Mesh::from(shape::Plane { size: 10.0 })),
+            mesh: meshes.add(Mesh::from(shape::Plane { size: 5.0 })),
             material: materials.add(Color::rgb(0.1, 0.2, 0.1).into()),
             ..Default::default()
         })
         // cube
         .spawn(PbrBundle {
-            mesh: meshes.add(Mesh::from(shape::Cuboid::new(1.0, 1.0, 1.0))),
+            mesh: meshes.add(Mesh::from(shape::Cube { size: 1.0 })),
             material: materials.add(Color::rgb(0.5, 0.4, 0.3).into()),
-            transform: Transform::from_translation(Vec3::new(0.0, 1.0, 0.0)),
+            transform: Transform::from_translation(Vec3::new(0.0, 0.5, 0.0)),
             ..Default::default()
         })
         // sphere
@@ -53,7 +53,7 @@ fn setup(
         })
         // camera
         .spawn(Camera3dBundle {
-            transform: Transform::from_translation(Vec3::new(-3.0, 5.0, 8.0))
+            transform: Transform::from_translation(Vec3::new(-2.0, 2.5, 5.0))
                 .looking_at(Vec3::default(), Vec3::unit_y()),
             ..Default::default()
         });
