+++ examples/shader/shader_custom_material.rs
@@ -84,7 +84,7 @@ fn setup(
     commands
         // cube
         .spawn(MeshBundle {
-            mesh: meshes.add(Mesh::from(shape::Cuboid::new(1.0, 1.0, 1.0))),
+            mesh: meshes.add(Mesh::from(shape::Cube { size: 2.0 })),
             render_pipelines: RenderPipelines::from_pipelines(vec![RenderPipeline::new(
                 pipeline_handle,
             )]),
